#17 Mapのスクリプト分離およびemitによる親ココンポーネントへのデータ送信
